Fixed build failures - Removed references to deleted files - If we link staticly, check for static zlib   - This should fix the problem with 'no -lz found' link error - Fixed build failure on window (Patch from Wlad) - Fixed build problem with federatedx when using -Werror
BUILD/Makefile.am:
  Remove removed file
config/ac-macros/zlib.m4:
  If we compile with --all-static, test that we have a static libz
libmysqld/CMakeLists.txt:
  Fix for build error on windows
mysql-test/suite/pbxt/r/key_cache.result:
  Updated result
mysql-test/suite/pbxt/t/key_cache.test:
  Fixed not updated test case
sql/CMakeLists.txt:
  Fix for build error on windows
storage/federatedx/Makefile.am:
  Don't use CFLAGS to compile C++ programs
storage/pbxt/src/lock_xt.cc:
  Fixed compiler warning about using uninitialized b2
storage/xtradb/buf/buf0buf.c:
  Fixed wrong printf
storage/xtradb/srv/srv0srv.c:
  Fixed assignment of different width and test with different sign/unsigned

+++ storage/xtradb/buf/buf0buf.c
@@ -1076,8 +1076,9 @@ init_again:
 			"InnoDB: Logical offset (blocks)         : %ld (%#lx)\n",
 				(byte*)chunk->mem + shm_info->frame_offset,
 				chunk->blocks[0].frame, frame,
-				phys_offset, phys_offset, logi_offset, logi_offset,
-				blocks_offset, blocks_offset);
+				phys_offset, phys_offset,
+                                (long) logi_offset, (long) logi_offset,
+				(long) blocks_offset, (long) blocks_offset);
 		} else {
 			fprintf(stderr,
 			"InnoDB: Buffer pool in the shared memory segment can be used as it is.\n");
